requirements.txt
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.6
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-js-asset==1.2.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
newrelic==6.8.0.163
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pafy==0.5.5
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pyglet==1.5.18
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
youtube-dl==2021.6.6
zipp==3.5.0
